Adding support for experimental 'metadata' block.
This adds support for putting top level fields into a 'metadata' block for organization purposes.

It also ships a 'conf' field to the publish endpoint containing the full contents of the `massdriver.yaml` file.

The goal is that the CLI will not need to be kept aware of changes to the spec and will simply ship the config off to the pkg manager.

This will make it easier to make changes to the structure of massdriver YAML files since it doesnt have to be changed in the pkg manager and the CLI.

TODOs:

* [ ] remove TODOs / notes
* [ ] Fix validate schema logic (bundle.LintSchema)
* [ ] Move JSON Schemas to gh json-schemas repo, pull for CLI & backend
* [ ] Add code to 'upgrade' schema definitions
* [ ] Remove legacy