Update generate command to use a docker container
Use a docker container to run the schema generation so
mix and other tools are not required locally. This relies on the
dev image in massdriver and will build it if not available.

Update the go generate command to specify a version which allows the
gen to run without having to modify the go.mod file